Fix 64-bit build conflicts and finalize audio gyroid fixes
- Fixed conflicting forward declarations in system.c for
  __Load_Wave and Nas_StartDma. They now correctly use uintptr_t
  for addresses, matching their 64-bit aware definitions.
- Resolved type conflicts in rspsim.h by updating RspStart and
  RspStart2 to use Acmd*.
- Verified that all jaudio_NES modules now use pointer-sized types
  (uintptr_t, OSMesg) for addresses and message passing.
- These changes resolve build failures on 64-bit platforms and
  fix the gyroid "stuck" animation and stopping/restarting crashes.

+++ src/static/jaudio_NES/internal/system.c
@@ -658,7 +658,7 @@ static void* FASTDMA_BUFFER = NULL;
 Na_DmaProc NA_DMA_PROC = Z_osEPiStartDma;
 Na_SyncProc NA_SYNC_PROC = Nas_GetSyncDummy;
 
-static s32 Nas_StartDma(OSIoMesg* ioMsg, s32 priority, s32 direction, u32 device_addr, void* dram_addr, u32 size,
+static s32 Nas_StartDma(OSIoMesg* ioMsg, s32 priority, s32 direction, uintptr_t device_addr, void* dram_addr, u32 size,
                         OSMesgQueue* mq, s32 medium, s8* dma_type);
 s32 Nas_BankOfsToAddr(s32 bank_id, u8* ctrl_p, WaveMedia* wave_media, s32 async);
 
@@ -670,7 +670,7 @@ static s32 __Kill_Bank(s32 bank_id);
 static ArcHeader* __Get_ArcHeader(s32 table_type);
 static s32 __Nas_StartSeq(s32 group_idx, s32 seq_id, s32 param);
 static u8* __Load_Bank(s32 table_type, s32 id, s32* did_alloc);
-static u32 __Load_Wave(s32 wave_id, u32* medium, s32 no_load);
+static uintptr_t __Load_Wave(s32 wave_id, u32* medium, s32 no_load);
 static void* __Check_Cache(s32 table_type, s32 id);
 #ifdef TARGET_PC
 static void __WaveTouch(wtstr* wavetouch_str, uintptr_t ram_addr, WaveMedia* wave_media);
